+++ src/main/java/kr/hhplus/be/server/domain/payment/PaymentRepository.java
@@ -9,6 +9,4 @@ public interface PaymentRepository {
 	Optional<Payment> findByReservationId(UUID reservationId);
 
 	void deleteAll();
-
-	Optional<Payment> findByReservationIdForUpdate(UUID reservationId);
 }
