+++ src/main/java/kr/hhplus/be/server/domain/seat/SeatRepository.java
@@ -8,12 +8,8 @@ public interface SeatRepository {
 
 	Seat save(Seat seat);
 	List<Seat> findAvailableSeats(UUID concertId, UUID concertDateId);
-
 	Optional<Seat> findById(UUID seatId);
+	Optional<Seat> findBySeatIdAndConcertDateId(UUID seatId, UUID concertDateId);
 
 	void deleteAll();
-
-	int updateStatusReserved(UUID seatId);
-
-	Optional<Seat> findBySeatIdAndConcertDateId(UUID seatId, UUID concertDateId);
 }
