+++ src/main/java/kr/hhplus/be/server/domain/user/UserRepository.java
@@ -6,7 +6,6 @@ import java.util.UUID;
 public interface UserRepository {
 	User save(User user);
 	Optional<User> findById(UUID userId);
-	Optional<User> findByIdForUpdate(UUID userId);
 
 	boolean existsById(UUID userId);
