+++ src/main/java/kr/hhplus/be/server/infrastructure/persistence/payment/jpa/JpaPaymentRepository.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.infrastructure.persistence.payment.jpa;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaPaymentRepository extends JpaRepository<PaymentEntity, String> {
+	Optional<PaymentEntity> findByReservationId(String reservationId);
+}
