+++ src/main/java/kr/hhplus/be/server/infrastructure/persistence/seat/SeatJpaGateway.java
@@ -23,11 +23,6 @@ public class SeatJpaGateway implements SeatRepository {
 	}
 
 	@Override
-	public int updateStatusReserved(UUID seatId) {
-		return jpaSeatRepository.updateStatusReserved(seatId.toString());
-	}
-
-	@Override
 	public Optional<Seat> findBySeatIdAndConcertDateId(UUID seatId, UUID concertDateId) {
 		return jpaSeatRepository.findBySeatIdAndConcertDateId(seatId.toString(), concertDateId.toString())
 			.map(SeatEntity::toDomain);
