+++ src/main/java/kr/hhplus/be/server/usecase/user/UserService.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Service;
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.framework.exception.CustomException;
 import kr.hhplus.be.server.framework.exception.ErrorCode;
-import kr.hhplus.be.server.infrastructure.persistence.lock.DistributedLockManager;
 import kr.hhplus.be.server.infrastructure.persistence.user.UserManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,10 +18,7 @@ import lombok.extern.slf4j.Slf4j;
 public class UserService {
 
 	private static final BigDecimal MIN_CHARGE_POINT = BigDecimal.valueOf(1000);
-	private static final String LOCK_KEY = "user:";
-
 	private final UserManager userManager;
-	private final DistributedLockManager distributedLockManager;
 
 	public User getUser(UUID userId) throws CustomException {
 		return userManager.getUser(userId);
@@ -34,13 +30,7 @@ public class UserService {
 			throw new CustomException(ErrorCode.NOT_ENOUGH_MIN_CHARGE_POINT);
 		}
 
-		String lockKey = LOCK_KEY + userId;
-		// user:{userId} 락 획득 후 포인트 충전 트랜잭션 수행
-		User user = distributedLockManager.executeWithLockHasReturn(
-			lockKey,
-			() -> userManager.chargePoint(userId, point)
-		);
-
+		User user = userManager.chargePoint(userId, point);
 		log.info("유저 포인트 충전: USER_ID - {}, CHARGE_POINT - {}, AFTER_POINT - {}", userId, point, user.amount());
 		return user;
 	}
